Simple JustMonika preview
Add a custom view as a preview to the main window. We take it directly
from the screensaver plugin which is embedded into the application.
That's not a library so we have to load it explicitly on startup.
While we're here, mark the view as "designable" so that we have a nice
preview in Interface Builder.