Display static Monika
Add original Monika background image into assets. Alternatively we could
have just placed it there manually, but let's use 'proper' means. I want
to study the Apple way of doing things.

Override awakeFromNib to setup the view in the preview application. This
is like usual for interfaces created in Interface Builder.

It turns out that it's not enough to just slap an IB_DESIGNABLE onto the
view. While its drawing mechanisms are called as expected, the view is
initalized differently. There is a special method that we need to call.

With this the view behaves identically in preview app, in designed UI in
Interface Builder, in screen saver preview, and as a screen saver. It's
even 'functional', in a way.