Introduce sublayers
We'll need to have quite a few moving parts here so we'll organize
them into sublayers. For example, the highlight is implemented by
a similar bitmap with varying opacity. We don't have opacity animation
now, but here's a bitmap.

Sublayers seem to be tricky. For now I've hacked their positioning
via layouts, but I think there could be a better way. To use constraints
we need to explicitly import the QuartzCore framework which we have been
using implicitly via ScreenSaver import up until this point.