Settings prank
Initial implementation of a nice prank: provide a settings button
which gets permanently disabled, similar to fast-forward in the game.

We have to put user interface into a XIB/NIB, load it manually, have
some weird 'file owner' thingie set up. Plus, we have to store settings
in a special NSUserDefaults instance because this is a Screen Saver.

However, we get a fun part of looking for the settings button in the
macOS Screen Saver settings dialog.

The issue here is that I have no idea how to reset the setting,
other than programmatically. Uninstalling the screen saver does not
work, the settings are still persisted somewhere.