Space background
Implement moving stars and nebula background. That's "mask_2" and
"mask_3" textures from the game and a simple movement animation.
It turned out to be way easier than I thought.

Loading the textures produces the following warning:

    libpng error: iCCP: known incorrect sRGB profile

I guess that's fine. It seems that the texture file has incorrect data
in it. Well, not much that we can do about it. I want to keep the same
image data so I won't repackage it.