Specify UV transformation in fragment shader
We have so many hardcoded instances of "1280" and "720" in there that
it does not have any sense to pretend that we support texture of other
size. Get rid of UV transformation matrix setup in C and just apply it
directly in GLSL code. It's not going to change.

Note that MVP matrix *does* change depending on the viewport. Though
we could probably move the vertex coordinates into the vertex shader
since we depend on them being equivalent to UV coordinates which are
in turn expected to be the same as image texel coordinates.