Offset adjustment UI in preview
Add a simple auxiliary window to control adjustments of offset of the
cloud texture. Expose the controls via temporary undocumented function.
Bind the slider value to displayed text edits and Monika configuration.

Note that we have to add our drawing timer to the event tracking run
loop mode because default mode does not fire when a slider is tracked.