More robust resource loading API
Yes, this is a bit of idiotic API, but it frees me from having to update
the shader buffer length so that we can read them. Now the caller is
expected to provide a malloc-allocated buffer (or NULL), the buffer will
be resized appropriately with realloc(), and then you should free() it.
Useful size is returned. Zero in case of errors.

Ideally I would like to bake in the shader source code into the library,
but I have no idea how to make this "resource embedding" portable. There
is some obscure GCC feature that would allow baking in resources, but on
macOS NSBundle can bake in only strings.