Use rectangular textures
What if I told you that you can stop doing all this fuckery with square
power-of-two sizes, padding, normalization, and just use pixel coords
directly? Wouldn't that be nice? It turns out that OpenGL can do that
on modern hardware (though yeah, historically it *was* incapable of
handling non-square, non-power-of-two, non-normalized textures).

Anyhow, switch to GL_TEXTURE_RECTANGLE type for textures. This means
that we cannot ever use built-in mipmapping, cannot use GL_REPEAT
wrapping (and have to do it ourselves when needed), do not need all
that emulation of GL_CLAMP_TO_EDGE wrapping as OpenGL does that now,
and we need to use a different sampler with simpler coordinates.

Overall, this means less code and less magic. I'd take it any day.
It's a shame I did not think about it earlier.

Let's hope this is supported in OpenGL ES (cause it likes to restrict
operations just to be a dick) and WebGL (which is OpenGL ES for Web).

And yep, since we do not use mipmaps, the image still looks crappy with
small viewports. So we may need to use framebuffers after all to do some
postprocessing.