Blur output for better preview
My university course in signal and image processing was not for naught:
I still had an intuitive understanding that the image has to be blurred
before downsampling to reduce the effect of high-frequency components.

Well, do just that with a simple Gaussian blur filter. I'm not sure what
radius to use, so add some UI to tweak it visually.