Add thumbnail image
This isn't documented anywhere, but screen savers can have custom
thumbnails if you provide an image named "thumbnail" in resources.
Usually this is two images "thumnail.png" (90x56) and "thumbnail@2x.png"
(180x116) which are intended to support both non-Retina and Retina
displays. These are combined into a single "thumbnail.tiff" file with
two pages when building. NSImage knows how to deal with that.

The original image here is "thumbnail@2x.png", the 1x size is just a
downscaled version of it.

Unfortunately, the code that processes thumbs is broken. It does some
ridiculous things which seem to amount to resizing the image to 90x56,
then upscaling the result to 180x116 to support Retina, and then add
that fancy frame. Somewhere in the middle the code fails to clip the
image (however, according to disassembly the clipping should be there).

Well... Apple can't be arsed to fix that code, that does not bring them
any more monies. So I guess fuck it then? I'm going to deal with this
later. I have an amazing idea for what to do with this thing. But for
now let the thumbnail look like shit. Blame Tim Cook, idk.