Clarify licensing issues
Usually I opt for using GPL for standalone software that I publish and
permissive licenses for libraries. Given that I would like to use some
questionable approaches in the screen saver implementation, I will use
GPL for it so that the users can always ensure that the code they run
does not do any shady things. The OpenGL library may be reused in other
contexts so it is licensed under Apache Software License 2.0 because it
is compatible with GPLv3 and does not grant trademark rights which is
important for redistributed DDLC assets.

Note that DDLC assets are a separate thing and they are mostly covered
by Team Salvato IP Guidelines. They can technically be swapped out for
redistribution if necessary.

Replace all Xcode's default boilerplate comments with more precise
information on the license. Use SPDX machine-readable identifiers.