Sync NSUserDefaults
It's a great mystery where Screen Saver framework actually keeps its
preferences. They don't show up in "defaults" output. There seems to
be a file in ~/Library/Preferences/ByHost which sometimes gets created
but removing it or changing values in it does not affect the screen
saver settings, as if the values are cached somewhere else.

Furthermore, ScreenSaverDefaults seem to require usage of "synchronize"
despite this method being marked as legacy. I don't really want to dig
into disassembly right now (though I'm definitely capable of doing it).

I've considered making this settings lockdown to be temporary by storing
some "disabled since" time in defaults and reenabling the settings back
after say 1 week or 1 month. However, that's too much of a bother and
may actually spoil the joke.

Currently [JustMonikaSettings reset] is never called, but I occasionally
use it during debugging. I guess a good moment to call it will be when
automatic updates are implemented. Then with a new version of screen
saver we can actually enable the button back (but I guess we can take
an option to leave it disabled permanently too).