Add Sparkle via Cocoapods
They recommend running with "use_frameworks!" but libpng cannot support
that (and that pod does not really seem to be supported, but it's kinda
official so I wouldn't like to take care of that myself). Anyhow, we can
embed the framework manually since Sparkle just distributes binaries.

Update the search paths too to include "$(inherited)" paths as suggested
by CocoaPods.