Integrate Sparkle autoupdater
Add a couple of special Info.plist keys for the updater to work. They
include my public key that will be used to sign the updates, and a dummy
reference to appcast.xml file. Since the project repo is not public yet
I'm currently using a secret Gist to host the file for public access.
The file currently points to a fake release of version "3" which will
trigger an update request for Sparkle.

We have to disable App Sandbox for the preview app because Sparkle v1
does not support sandboxing (to the point of throwing exceptions if it
is enabled). The next major version will support it but we're not really
concerned with that since Screen Saver plugins are not sandboxed at all,
or so it seems.

All updating stuff is managed by a separate object which also acts as
a delegate to Sparkle. We need to tweak the behavior here and there to
ensure that it behaves nicely in the screen saver envionment. For one,
we cannot restart the application after update and we should not display
alerts when a screen saver is running. Well, you can read the comments.

Note that we disable sending system reports explicitly, just in case
Sparkle decides to change their default of not sending them. I don't
want to collect any information on the users, so no phoning home.

Overall, Sparkle is quite pleasant to work with. I give them respect.