Notify the user about updates
Currently we check for updates all the time (if allowed) but we show
an alert only when the user opens System Preferences and looks at the
screen saver preview. This means that most of the time--when running
an actual screen saver--the user will not even know about an update.

Let's be a little more persistent and show a notification when we see an
update available. This will be done only when running the actual screen
saver. Mostly because System Preferences is not allowed to send user
notifications. I'm actually surprised that ScreenSaverEngine is allowed
to send notifications--and is granted permissions for that without any
user interaction! Or so it seems to be on my macOS 10.14.

Make sure that we are as discreet as possible and avoid nagging the
user too much for regular updates. Normally, a notification should be
shown exactly once. Make an exception for critical updates--well, they
*are* critical for a reason--and for them show a new notification on
every launch. The user still has an option to disable notifications
entirely in System Preferences.