Use a custom alert category
Well... initially I wanted to see how far can I take this since we're
doing notifications now. I have noticed a weird behavior when clicking
the notification instead of dragging it away. It kept bothering me for
a while. And then I've came up with a decent approach to exploit this.
Read the comment.