Use designated initializer in NIBs
I don't like the code duplication from start and finally I have found
a way to avoid it. Let's call the designated initializer, but provide
a way for the Interface Builder to set the preview setting. It turned
out to be easier that I thought.

Now I can test both modes, and we can use ScreenSaverView's native
timers for animation! But not for rendering so we still run our own
NSTimer. However, it's now safe to call "startAnimation".

This means we can drop that custom category from JustMonikaView which
is used to get direct access to JustMonikaGLView instance and move
that header into project headers.