Fake Monika for Interface Builder
Since I have ported this to OpenGL I have been experiencing issues with
Interface Builder preview. Maybe that's because my code is shit (most
likely), or because Apple code is shit (also likely), but Interface
Builder says that it cannot render an image in 200 ms so no previews.

Previously I was able to drop into a debugger, but after updating Xcode
this is no longer possible because "My Mac cannot run i386 executables".
Apparently, Interface Builder builds previews as 32-bit applications...
and refuses to launch them. Idiocy.

Furthermore, Interface Builder complains about "200 ms" for whatever
reason, be it either time issues, or an exception in the drawing code.
I have no way to tell that without a debugger.

So replace OpenGL Monika with a Potyomkin poster of her :) "No, much
better! It's a drawing of a key!" And apparenly we can't use original
resources because they too cannot be loaded in 200 ms. Okay, ffs, you
win, piece of software shit. Take that and stop bothering me.

(And there are some changes in XIB because I just discovered them!)