Support older NSUserNotification API
Well... why not?

Downgrade our target API to 10.12 -- that's a minimum required for some
other stuff, like timers and image loading. Well, only Sierra currently
has "extended" support so there should be no issues with not allowing
anything earlier. I could probably go back as far as 10.8 or something,
but I'm too lazy. The current only user runs 10.14 anyway.

Don't forget to update CocoaPods spec too so that the dependencies are
built for 10.12 as well. And link to UserNotifications.framework weakly.
This means that we have to check API availability. I have also added
some magic macros to mark when each API appeared and went away.

Now, the notification API is mostly reasonable. For one, it does not
launch your application if it's closed and you didn't ask for relaunch.
Therefore we provide no action tricks. Even we did, unfortunately the
actions will not be delivered to anyone -- because the screen saver
process has already exited -- so we cannot launch a new one.

What is absurd is that those fucktards in Apple actually track what API
your application used and deny usage of the different API! Therefore
you have to change bundle ID of the preview app to get it display the
NSUserNotifications. And you have to manually remove ScreenSaverEngine
from user defaults via "defaults" and reboot for it to have an effect.