Round edges of thumbnails
Apple's shit code for thumbnail handling is a story for the other day.
I'll get back to improving the thumbnails later. For now, apply the
rounded corner mask because Apple is too lazy to apply it themselves
(despite actually having it, and the code that transforms our thumb
into whatever god-awful-resampled version they show for third-party
screensavers).