Replace thumbnail with high-quality version
For some reason screen saver thumbnails look like shit. I mean, it looks
as if Apple first rescales the thumbnail to 90x56, then adds a frame,
and then upscales that back to 180x116 for display. Retina? What's that?

I have even shoveled that preference pane into IDA in order to learn
what the fuck are they doing there, but I wasn't patient enough.

Therefore, take the matter into our own hands. Exploit the fact that we
have access to the view hierarchy, locate the view that displays the
thumbnail, and update its "image" property so that the thumnail actually
uses HiDPI images.

It's definitely not suitable for AppStore, I guess, but I don't care.