Tweak mask position
The mask seems to be offset by a whole pixel in Apple's built-in screen
savers. I admit that I did not look too deeply in the original framing
implementation. Well, tweak it so that our result looks similar.

The mask seems to still need some subpixel adjustments, but it's good
enough for now.