Animate thumbnail replacement
Instead of replacing the thumbnail instantly, play a little bit of white
noise in the frame before making the switch. I admit that the approach
is somewhat suboptimal, but it's good enough for now.

Also, when filtering potential victims it's okay to replace those which
are only partially visible. Usually it's only a couple of pixes which
are missing so it would be a waste to not replace them too if chances
say that we should.