Handle switching between tabs
If the user switches to "Desktop" tab and back, Monika goes on with her
screen saver replacement agenda, but she replaces them with some random
other screen saver, not with herself. Also, when she kills everyone then
only that random screen saver survives. This is not what we want.

Turns out that NSCollectionView may repurpose its child views so after
we do a "reloadData" call then the view may get replaces with something
else. I guess it's also possible if Monika stays out of sight for long.

Instead of carrying around an NSView as a proxy for Monika's name and
image, carry around those directly. This fixes the faulty behavior.

There's also another one when we switch to a different settings pane.
Monika seems to be still running in this case, and when the user
switches back to screen saver settings then she gets disoriented a bit
and again replaces things at random.

This is probably because the collection data source gets recreated, with
the old Monika image still in use. This means that Monika can replace
even herself!