Tweak thumbnail improvements
Move dispatch_once() call a little bit deeper. The thing is that we have
to call the thumbnail improvement code on every start of the animation
because the screen saver list may have been recreated without our bundle
being reloaded. This routinely happens when the user switches to some
other preference panel and then goes back to screen savers.

However, we need to initiate the Screen Saver Instrumentality Project
only once since the windows and views are not recreated. We can expect
the NSCollectionView to stay there, but its subviews may change.