Identify Monika by name
Not by her image. This is because the name is not going to change
between reloads, but the image may get reset back to shitty thumbnail.
It's more reliable to use the name. The name will not change even if the
user changes system language. That happens only when the application
restarts.

There are still some funny behaviors that I observed when testing all
the tab and window switching, but I'm too lazy to iron them out now.
This is probably good enough for a one-off joke.