Start publishing appcast.xml on GitHub
Move it here from Gist and expect it to be available on the master
branch. Every update to this file on the master branch may trigger
software update.

Currently we're using a dummy package hosted on Gist for testing.
Later I'll update the URL in appcast to point to GitHub releases.