Fix embedded code signing
You cannot update the signature of Autoupdate.app after its containing
framework has been signed. This leads to a broken signature which is not
immediately apparent during development, but makes Gatekeeper scream
APPLICATION IS DAMAGED YOU SHOULD MOVE IT TO TRASH AND KILL IT WITH FIRE

Disabled "code sign on copy" in Embed Frameworks stage and sign them
manually with "--force --deep" so that Autoupdate.app is resigned.

I'm going to rebuild v1.1 binaries with this change, but without making
an official release. Nobody will notice a thing, but appcast.xml needs
to be updated.