React to character file removal
This is on my gimmick list. Let's actually check if the user has been
keeping the character file safe. If they are assholes then don't draw
Monika on the screen. It's actually very easy to do.

For testing purposes, expose JustMonikaGLView to the preview app and
add some UI controls to switch the display on and off. While we're here,
setup proper layout constraints in the adjustment helper window.

Note that removing the character file compromises code signature:

    $ codesign --verbose --verify --deep --strict JustMonikaPreview.app
    JustMonikaPreview.app: a sealed resource is missing or invalid
    In subcomponent: JustMonikaPreview.app/Contents/PlugIns/JustMonika.saver
    file missing: JustMonikaPreview.app/Contents/PlugIns/JustMonika.saver/Contents/Resources/monika.chr

so I'm not sure if it will work on Catalina but my Mojave has no issues
with that once the screen saver has been installed.