Describe the release process
I have dug out an old draft of the release process and updated it. Now
we don't need to provide Sparkle stuff. However, since then I have
obtained a Developer ID certificate from Apple so it's possible to
notarize the screen saver. Unfortunately, Xcode does not support this
kind of bundles so the process needs to be performed manually. It may be
better to write a script for this, but I don't imaging doing this all
too often, so this document should be a fine guide for now.