Fix viewport size computation on Catalina
I have upgraded to macOS 10.15 Catalina (at last) and have noticed that
Monika is a bit small--to be exact, a quarter of what she should be.

This is because I don't read the documentation for NSView's "bounds"
property and have not noticed that it might not correspond exactly to
the device coordinate system. Start using the "convertSizeToBacking"
to convert Cocoa's size to device pixel size--the one we need to use
with OpenGL.

This fixes Monika's size on Catalina. There are no other major issues
with core functionality, thankfully, but I guess there may be issues
with some of the gimmicks.