Show "options" sheet only once
Unfortunately, we can't disable the button right away since macOS 10.15
due to the architectural changes. The preferences panel still caches the
result of "hasConfigureSheet" so the button is not disabled until the
window has reopened.

So after the sheet has been disabled, start returning nil for it so that
it does not show again, even if the user presses the button again. It's
not very subtle but at least it's better than showing it over and over.