Decode PNG textures with Core Graphics
Instead of relying on third-party libpng as a mandatory dependency
of JustMonikaGL, let's use Core Graphics system framework. It's a bit
clunky to use, but this should simplify the build system which can be
liberated from CocoaPods.

Note that we need to be extra-careful to use proper measurements of the
image as well as a proper format. OpenGL expects images to be in RGBA
format, with top line coming first in the buffer.